Split add event page into 3: - Select Venue - Select Sports - Customize Event
This was to facillitate a dynamic list of sports and dynamic list of venues

Made venue list page a recycler view that automatically listens to and adds
venues from the database

Sports list gets the list of sports passed on from the venue page's chosen venue

In customize event page the venue and sport field are fixed.

Added finish to the create event's transition back to main activity so that one cannot go back as going back
would cause issues with database writing.

TODO:
- Data validation
- Integration with User table to link events to creators
- Forgot to add "ArrayList<Integer> attendees" field to events class
- Need to modularize venue recycle view a bit so that it can be reused in other places with
  minimal code duplication

- Make create event page useful for modifying event
- Other stuff I haven't thought of yet LOL